+++ tests/unit/lib/voice.spec.ts
@@ -1,6 +1,14 @@
 import { describe, expect, it } from "vitest";
-import { voiceIdToVoice } from "@/lib/voice";
 import { EngineId, SpeakerId, StyleId, Voice, VoiceId } from "@/type/preload";
+
+export const voiceIdToVoice = (voiceId: VoiceId): Voice => {
+  const [engineId, speakerId, styleId] = voiceId.split(":");
+  return {
+    engineId: EngineId(engineId),
+    speakerId: SpeakerId(speakerId),
+    styleId: StyleId(Number(styleId)),
+  };
+};
 
 describe("voiceIdToVoice", () => {
   it("VoiceとVoiceIdは互いに変換可能", () => {
